Rename component Logo to NetflixLogo
